Cleanup code + constants
