+++ packages/mdx/src/plugins/rehype/rehypeSyntaxHighlighting.ts
@@ -1,51 +1,40 @@
 import type { Element, Root } from 'hast';
 import { toString } from 'hast-util-to-string';
-import { createHighlighter, type Highlighter, type BuiltinTheme, type BundledTheme } from 'shiki';
+import { createHighlighter, type Highlighter } from 'shiki';
 import type { Plugin } from 'unified';
 import { visit } from 'unist-util-visit';
 
 import {
+  type ShikiLang,
+  type ShikiTheme,
+  shikiColorReplacements,
   DEFAULT_LANG_ALIASES,
   SHIKI_THEMES,
   UNIQUE_LANGS,
   DEFAULT_LANG,
-  type ShikiLang,
-  type ShikiTheme,
+  DEFAULT_DARK_THEME,
+  DEFAULT_LIGHT_THEME,
+  DEFAULT_THEMES,
 } from './shiki-constants.js';
-
-const shikiColorReplacements: Partial<Record<BundledTheme, string | Record<string, string>>> = {
-  'dark-plus': {
-    '#1e1e1e': 'transparent',
-    '#569cd6': '#9cdcfe',
-    '#c8c8c8': '#f3f7f6',
-    '#d4d4d4': '#f3f7f6',
-  },
-  'github-light-default': {
-    '#fff': 'transparent',
-    '#ffffff': 'transparent',
-  },
-};
+import {
+  getLanguage,
+  getLinesToHighlight,
+  lineHighlightPattern,
+  LINE_HIGHLIGHT_CLASS,
+} from './utils.js';
 
 export type RehypeSyntaxHighlightingOptions = {
-  theme?: BuiltinTheme;
-  themes?: Record<'light' | 'dark', BuiltinTheme>;
+  theme?: ShikiTheme;
+  themes?: Record<'light' | 'dark', ShikiTheme>;
   codeStyling?: 'dark' | 'system';
 };
-
-const lineHighlightPattern = /\{(.*?)\}/;
-
-function classNameOrEmptyArray(element: Element): string[] {
-  const className = element.properties.className;
-  if (Array.isArray(className) && className.every((el) => typeof el === 'string')) return className;
-  return [];
-}
 
 let highlighterPromise: Promise<Highlighter> | null = null;
 
 async function getHighlighter(): Promise<Highlighter> {
   if (!highlighterPromise) {
     highlighterPromise = createHighlighter({
-      themes: ['github-light-default', 'dark-plus'],
+      themes: DEFAULT_THEMES,
       langs: UNIQUE_LANGS,
     });
   }
@@ -57,14 +46,18 @@ export const rehypeSyntaxHighlighting: Plugin<[RehypeSyntaxHighlightingOptions?]
 ) => {
   return async (tree) => {
     const highlighter = await getHighlighter();
-    if (options.theme) {
+    if (options.theme && !DEFAULT_THEMES.includes(options.theme)) {
       await highlighter.loadTheme(options.theme);
     }
     if (options.themes) {
-      await Promise.all([
-        highlighter.loadTheme(options.themes.dark),
-        highlighter.loadTheme(options.themes.light),
-      ]);
+      const promises: Promise<void>[] = [];
+      if (!DEFAULT_THEMES.includes(options.themes.dark)) {
+        promises.push(highlighter.loadTheme(options.themes.dark));
+      }
+      if (!DEFAULT_THEMES.includes(options.themes.light)) {
+        promises.push(highlighter.loadTheme(options.themes.light));
+      }
+      await Promise.all(promises);
     }
 
     visit(tree, 'element', (node, index, parent) => {
@@ -106,8 +99,8 @@ export const rehypeSyntaxHighlighting: Plugin<[RehypeSyntaxHighlightingOptions?]
             light:
               options.themes?.light ??
               options.theme ??
-              (options.codeStyling === 'dark' ? 'dark-plus' : 'github-light-default'),
-            dark: options.themes?.dark ?? options.theme ?? 'dark-plus',
+              (options.codeStyling === 'dark' ? DEFAULT_DARK_THEME : DEFAULT_LIGHT_THEME),
+            dark: options.themes?.dark ?? options.theme ?? DEFAULT_DARK_THEME,
           },
           colorReplacements: shikiColorReplacements,
           tabindex: false,
@@ -134,9 +127,9 @@ export const rehypeSyntaxHighlighting: Plugin<[RehypeSyntaxHighlightingOptions?]
           lineNumber++;
           if (linesToHighlight.includes(lineNumber)) {
             if (typeof span.properties.class === 'string') {
-              span.properties.class += ' line-highlight';
+              span.properties.class += ' ' + LINE_HIGHLIGHT_CLASS;
             } else {
-              span.properties.class = [...span.properties.class, 'line-highlight'];
+              span.properties.class = [...span.properties.class, LINE_HIGHLIGHT_CLASS];
             }
           }
         });
@@ -164,51 +157,5 @@ export const rehypeSyntaxHighlighting: Plugin<[RehypeSyntaxHighlightingOptions?]
     });
   };
 };
-
-function getLanguage(node: Element, aliases: Record<string, ShikiLang>): ShikiLang | undefined {
-  const className = classNameOrEmptyArray(node);
-
-  for (const classListItem of className) {
-    if (classListItem.startsWith('language-')) {
-      const lang = classListItem.slice(9).toLowerCase();
-      if (lang) return aliases[lang];
-    }
-  }
-
-  return undefined;
-}
-
-function getLinesToHighlight(node: Element, maxLines: number): number[] {
-  const meta =
-    typeof node.data?.meta === 'string'
-      ? node.data.meta
-      : classNameOrEmptyArray(node).reduce((acc, item) => acc + ' ' + item, '');
-  if (!meta) return [];
-
-  const content = meta.match(lineHighlightPattern)?.[1]?.trim();
-  if (!content) return [];
-
-  const lineNumbers = new Set<number>();
-
-  content.split(',').forEach((part) => {
-    const [start, end] = part.split('-').map((num) => {
-      const trimmed = num.trim();
-      if (!/^\d+$/.test(trimmed)) return undefined;
-      const parsed = parseInt(trimmed, 10);
-      return parsed > maxLines ? maxLines : parsed;
-    });
-
-    if (!start) return;
-    const endLine = end ?? start;
-
-    if (endLine < start) return;
-    const max = Math.min(endLine, maxLines);
-    for (let i = start; i <= max; i++) {
-      lineNumbers.add(i);
-    }
-  });
-
-  return Array.from(lineNumbers).sort((a, b) => a - b);
-}
 
 export { UNIQUE_LANGS, DEFAULT_LANG_ALIASES, SHIKI_THEMES, ShikiLang, ShikiTheme };
